+++ lib/blocs/rpc_bloc.dart
@@ -1,7 +1,9 @@
 import 'dart:io';
 
 import 'package:bloc/bloc.dart';
+import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as shelf_io;
+import 'package:shelf_cors_headers/shelf_cors_headers.dart';
 import 'package:sovarpc/blocs/rpc_events.dart';
 import 'package:sovarpc/services/settings_yaml.dart';
 
@@ -66,10 +68,13 @@ class RpcBloc extends Bloc<RPCEvents, RpcState> {
         rpcAddress: address, ignoreMethods: iMethods, rpcRunnable: false));
   }
 
+  /// Added support for CORS (CORS_ALLOW_ALL_ORIGINS = True)
+  /// v1.2.4
   void _startServer(event, emit) async {
     try {
       var address = event.address;
       var ignoreMethods = event.ignoreMethods;
+      var whiteList = event.whiteList;
       var addressArray = address.split(":");
       var config = locatorRpc<SettingsYamlHandler>();
       await config.writeSet(SettingsKeys.ip, addressArray[0]);
@@ -80,10 +85,22 @@ class RpcBloc extends Bloc<RPCEvents, RpcState> {
         await rpcServer!.close(force: true);
         rpcServer = null;
       }
+
+      bool checker(String x) {
+        return true;
+      }
+
+      final pipeline = const Pipeline()
+          .addMiddleware(corsHeaders(headers: {
+            'Access-Control-Allow-Origin': '*',
+            'Access-Control-Allow-Methods': 'GET, POST, OPTIONS',
+            'Access-Control-Allow-Headers': 'Content-Type',
+          }, originChecker: checker))
+          .addHandler(
+              ServiceRPC(_repositories, ignoreMethods, whiteList).handler);
+
       rpcServer = await shelf_io.serve(
-          ServiceRPC(_repositories, ignoreMethods).handler,
-          addressArray[0],
-          int.parse(addressArray[1]));
+          pipeline, addressArray[0], int.parse(addressArray[1]));
 
       _debugBloc.add(AddStringDebug(
           "Start RPC at http://${rpcServer?.address.host}:${rpcServer?.port}",