+++ lib/database/database.dart
@@ -5,18 +5,18 @@ import 'package:drift/native.dart';
 import 'package:noso_dart/models/noso/address_object.dart';
 import 'package:path/path.dart' as p;
 
-import '../address_wallet.dart';
-import '../contact.dart';
+import '../../models/address_wallet.dart';
+import '../../models/contact.dart';
 import 'address.dart';
 import 'contact.dart';
 
 part 'database.g.dart';
 
 @DriftDatabase(tables: [Addresses, Contact])
-class MyDatabase extends _$MyDatabase {
+class LocalDatabase extends _$MyDatabase {
   final String pathDatabase;
 
-  MyDatabase(this.pathDatabase) : super(_openConnection(pathDatabase));
+  LocalDatabase(this.pathDatabase) : super(_openConnection(pathDatabase));
 
   @override
   MigrationStrategy get migration {