+++ lib/di.dart
@@ -5,14 +5,14 @@ import 'package:sovarpc/blocs/debug_rpc_bloc.dart';
 import 'package:sovarpc/blocs/noso_network_bloc.dart';
 import 'package:sovarpc/blocs/rpc_bloc.dart';
 import 'package:sovarpc/repository/file_repository.dart';
-import 'package:sovarpc/repository/repositories_rpc.dart';
-import 'package:sovarpc/services/file_service_rpc.dart';
-import 'package:sovarpc/services/settings_yaml.dart';
-import 'package:sovarpc/w_old/database/database.dart';
-import 'package:sovarpc/w_old/noso_network_service.dart';
 import 'package:sovarpc/repository/local_repository.dart';
 import 'package:sovarpc/repository/noso_network_repository.dart';
+import 'package:sovarpc/repository/repositories_rpc.dart';
+import 'package:sovarpc/services/file_service_rpc.dart';
+import 'package:sovarpc/services/noso_network_service.dart';
+import 'package:sovarpc/services/settings_yaml.dart';
 
+import 'database/database.dart';
 import 'models/log_level.dart';
 
 final GetIt locatorRpc = GetIt.instance;
@@ -26,15 +26,15 @@ Future<void> setupDiRPC(AppPath pathApp,
   locatorRpc.registerLazySingleton<LogLevel>(
       () => logLevel ?? LogLevel(level: "Debug"));
   locatorRpc
-      .registerLazySingleton<SettingsYamlHandler>(() => SettingsYamlHandler(pathApp));
+      .registerLazySingleton<SettingsYamlHandler>(() => SettingsYamlHandler());
 
   /// shared & drift(sql)
-  locatorRpc.registerLazySingleton<MyDatabase>(() => MyDatabase(pathApp));
+  locatorRpc.registerLazySingleton<LocalDatabase>(() => LocalDatabase(pathApp));
 
   /// repo && services
 
   locatorRpc.registerLazySingleton<RepositoriesRpc>(() => RepositoriesRpc(
-      localRepository: LocalRepository(locatorRpc<MyDatabase>()),
+      localRepository: LocalRepository(locatorRpc<LocalDatabase>()),
       networkRepository: NosoNetworkRepository(NosoNetworkService()),
       fileRepository: FileRepositoryRpc(
           FileServiceRpc(pathApp, nameFileSummary: "summaryRPC.psk")),