+++ lib/services/settings_yaml.dart
@@ -13,13 +13,13 @@ class SettingsKeys {
   static String logsLevel = "logsLevel";
   static String lastSeed = "lastSeed";
   static String rpcAddress = "rpcAddress";
+  static String whiteList = "whitelist";
+  static String verificationSeeds = "verification_seeds";
   static String defaultPaymentAddress = "defaultPaymentAddress";
 }
 
 class SettingsYamlHandler {
-  final String appPath;
-
-  SettingsYamlHandler(this.appPath);
+  SettingsYamlHandler();
 
   Future<List<String>> checkConfig() async {
     final File configFile = File(PathAppRpcUtil.rpcConfigFilePath);
@@ -35,6 +35,7 @@ class SettingsYamlHandler {
         settings[SettingsKeys.logsLevel] = "release";
         settings[SettingsKeys.defaultPaymentAddress] = "";
         settings[SettingsKeys.lastSeed] = "";
+        settings[SettingsKeys.whiteList] = "";
         settings.save();
         stdout.writeln(
             Pen().greenText('${PathAppRpcUtil.rpcConfigFilePath} created.'));
@@ -49,6 +50,7 @@ class SettingsYamlHandler {
       settings[SettingsKeys.logsLevel],
       settings[SettingsKeys.defaultPaymentAddress],
       settings[SettingsKeys.ignoreMethods],
+      settings[SettingsKeys.whiteList],
     ];
   }