+++ lib/utils/enum.dart
@@ -1,0 +1,7 @@
+enum StatusConnectNodes { connected, error, searchNode, sync, consensus }
+
+enum ApiStatus { connected, loading, error, result }
+
+enum ConsensusStatus { sync, error }
+
+enum InitialNodeAlgorithm { listenDefaultNodes, connectLastNode, listenUserNodes }