+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v.1.2.4
+
+- Added commands, --version, --testSeeds
+- Added support for manual installation of verified seeds
+- Add support white list IPs
+- Add support for CORS (CORS_ALLOW_ALL_ORIGINS = True)
+
 ## v.1.1.4
 
 - Fixing bugs and updating verification nodes